download logic

+++ dataplicity/remote_directory.py
@@ -323,29 +323,30 @@ class RemoteDirectory(object):
 
         device_path = write_device_path.encode("utf-8")
 
-        try:
-            # Write the data at the given offset
-            log.debug("writing %i bytes to %s", len(data), write_device_path)
-            with open(write_device_path, "r+b") as download_file:
-                download_file.seek(offset)
-                download_file.write(data)
-        except Exception as error:
-            # Error writing chunk, tell server
-            log.exception("error writing download")
-            client.send(
-                PacketType.write_remote_file_result,
-                id,
-                offset,
-                device_path,
-                2,
-                str(error).encode("utf-8", "replace"),
-            )
+        if data:
             try:
-                log.debug("removing %s", write_device_path)
-                os.remove(write_device_path)
-            except:
-                log.exception("failed to remove %r", write_device_path)
-            return
+                # Write the data at the given offset
+                log.debug("writing %i bytes to %s offset=%i", len(data), write_device_path, offset)
+                with open(write_device_path, "r+b") as download_file:
+                    download_file.seek(offset)
+                    download_file.write(data)
+            except Exception as error:
+                # Error writing chunk, tell server
+                log.exception("error writing download")
+                client.send(
+                    PacketType.write_remote_file_result,
+                    id,
+                    offset,
+                    device_path,
+                    2,
+                    str(error).encode("utf-8", "replace"),
+                )
+                try:
+                    log.debug("removing %s", write_device_path)
+                    os.remove(write_device_path)
+                except:
+                    log.exception("failed to remove %r", write_device_path)
+                return
 
         if final:
             try:
@@ -393,7 +394,7 @@ class RemoteDirectory(object):
                 raise TooManyCopies(
                     "Can't find unique filename for %s" % destination_path
                 )
-            destination_path = "%s/%s (%s)%s" % (dirname, basename, tries, ext)
+            destination_path = "%s/%s(%s)%s" % (dirname, basename, tries, ext)
 
         log.debug("copying %s to %s", device_path, destination_path)
         try:
